COMMON: Add PriorityQueue class
