+++ engines/ultima/shared/std/containers.h
@@ -87,8 +87,6 @@ public:
 		}
 	};
 public:
-	typedef T reference;
-	typedef const T const_reference;
 
 	vector() : Common::Array<T>() {}
 	vector(size_t newSize) : Common::Array<T>(newSize) {}
@@ -229,69 +227,6 @@ public:
 	reverse_iterator rend() {
 		return reverse_iterator(Common::List<T>::end());
 	}
-};
-
-/**
- * Queue ordered by a provided priority function
- * NOTE: Unlike in the C std library, we have to provde a comparitor that sorts
- * the array so that the smallest priority comes last
- */
-template <class _Ty, class _Container, class _Pr>
-class priority_queue {
-public:
-	priority_queue() : c(), comp() {}
-
-	explicit priority_queue(const _Pr &_Pred) : c(), comp(_Pred) {}
-
-	priority_queue(const _Pr &_Pred, const _Container &_Cont) : c(_Cont), comp(_Pred) {
-		make_heap(c.begin(), c.end(), comp);
-	}
-
-	template <class _InIt>
-	priority_queue(_InIt _First, _InIt _Last, const _Pr &_Pred, const _Container &_Cont) : c(_Cont), comp(_Pred) {
-		c.insert(c.end(), _First, _Last);
-		make_heap(c.begin(), c.end(), comp);
-	}
-
-	template <class _InIt>
-	priority_queue(_InIt _First, _InIt _Last) : c(_First, _Last), comp() {
-		make_heap(c.begin(), c.end(), comp);
-	}
-
-	template <class _InIt>
-	priority_queue(_InIt _First, _InIt _Last, const _Pr &_Pred) : c(_First, _Last), comp(_Pred) {
-		make_heap(c.begin(), c.end(), comp);
-	}
-
-	bool empty() const {
-		return c.empty();
-	}
-
-	size_t size() const {
-		return c.size();
-	}
-
-	typename _Container::const_reference top() const {
-		return c.back();
-	}
-
-	void push(const typename _Container::value_type &_Val) {
-		c.push_back(_Val);
-		Common::sort(c.begin(), c.end(), comp);
-	}
-
-	void pop() {
-		c.pop_back();
-	}
-
-	void swap(priority_queue &_Right) {
-		SWAP(c, _Right.c);
-		SWAP(comp, _Right.comp);
-	}
-
-protected:
-	_Container c;
-	_Pr comp;
 };
 
 } // End of namespace Std
